requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
openpyxl>=3.1.0
pandas>=2.0.0
dj-database-url>=2.1.0
django-cors-headers>=4.0.0
